requirements.txt
browser-use
langchain-anthropic
langchain-openai
python-dotenv
fastmcp 
mcp[cli]